Fixed routeing and added working tests
The nested routeing was not working so I've implemented it in a different way.  

I'm not sure it's the best way but it's the only one I could get to work with the new structure. I've added some working test which can be run with: 

PORT=9000 NO_SECURITY=1 npm test

This should make it easier to test this again if you want to refactor/try a different method.